libft

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: myavas <myavas@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/10 15:16:15 by myavas            #+#    #+#             */
-/*   Updated: 2024/11/12 13:40:15 by myavas           ###   ########.fr       */
+/*   Updated: 2024/11/14 17:36:45 by myavas           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while ((s[length] != '\0'))
